requirements.txt
flask==2.0.1
gunicorn==20.1.0
python-dotenv==0.19.0
web3==5.23.1